hmm error, gmm error
P(D|M) for an entire dataset is now calculated correctly. It used to be
the sum of P(x|M) where x is an element of the dataset, when it
should've been the product (sum of log probabilities). All models
should've been correct, but all probabilities were calculated
incorrectly.

Also, a minor error in HMMs in the backward algorithm for models without
an explicit end state.