added splitting of states
the chosen state will be split into two states with different
distributions. The distribution will be split according to its
implementation. The new state will have the same predecessors and
successors as the old one. The exiting edges of the predecessors will
be adjusted so that the edge probabilities add up to 1.